move MetabasePluginsConfig
