+++ frontend/src/metabase/embedding-sdk/types/plugins.ts
@@ -1,8 +1,9 @@
 import type { ReactNode } from "react";
 
 import type { DashCardMenuItem } from "metabase/dashboard/components/DashCard/DashCardMenu/DashCardMenu";
-import type { MetabaseQuestion } from "metabase/embedding-sdk/types/question";
 import type { ClickAction, ClickObject } from "metabase/visualizations/types";
+
+import type { MetabaseQuestion } from "./question";
 
 export type MetabaseDataPointObject = Pick<
   ClickObject,
